fix microservice one

+++ src/main/java/com/productService/productServiceAdd/Repository/ProductRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ProductRepository extends JpaRepository<Products,Long> {
+public interface ProductRepository extends JpaRepository<Products,Integer> {
 
   Optional<Products> findByProductName(String prosuctName);
 }
